Refactoring: function to create RegionalTaskRefactoring: moving excess code from R5RCore to other classes - Moved code from IsochroneBuilder to R5Process superclass, which will be shared with TravelTimeMatrixComputer and DetailedItineraryPlanner
